tests/requirements.txt
pytest==4.4.1
vcrpy==2.0.1